Extract HandleDefaultParamValuesPass::PrintExpression() so that it can be reused
